export styled components from separate file for Product Tile
